+++ mlxtend/evaluate/__init__.py
@@ -23,6 +23,8 @@ from .ttest import paired_ttest_5x2cv
 from .holdout import RandomHoldoutSplit
 from .holdout import PredefinedHoldoutSplit
 from .f_test import ftest
+from .f_test import combined_ftest_5x2cv
+
 
 
 __all__ = ["scoring", "confusion_matrix",
@@ -34,4 +36,4 @@ __all__ = ["scoring", "confusion_matrix",
            "paired_ttest_kfold_cv", "paired_ttest_5x2cv",
            "feature_importance_permutation",
            "RandomHoldoutSplit", "PredefinedHoldoutSplit",
-           "ftest"]
+           "ftest", "combined_ftest_5x2cv"]